requirements.txt
annotated-types==0.7.0
anyio==4.11.0
click>=8.1.7,<9
fastapi==0.118.0
python-multipart>=0.0.9,<1
h11==0.16.0
httptools==0.7.1
idna==3.11
pydantic==2.11.9
python-dotenv==1.1.1
PyYAML==6.0.3
sniffio==1.3.1
starlette==0.48.0
typing_extensions==4.15.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.1
websockets==15.0.1
sqlalchemy==2.0.43
psycopg[binary]==3.2.10
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
httpx==0.28.1
pytest==8.4.2
boto3>=1.35,<2
twilio>=9.0,<10
APScheduler>=3.10,<4.0